oracle should work

+++ lib/rucio/db/sqla/migrate_repo/versions/102efcf145f4_added_stuck_at_column_to_rules.py
@@ -1,38 +1,55 @@
-'''
- Copyright European Organization for Nuclear Research (CERN)
+# Copyright 2013-2019 CERN for the benefit of the ATLAS collaboration.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# Authors:
+#
+# - Martin Barisits <martin.barisits@cern.ch>, 2014
+# - Vincent Garonne <vincent.garonne@cern.ch>, 2017
+# - Mario Lassnig <mario.lassnig@cern.ch>, 2019
 
- Licensed under the Apache License, Version 2.0 (the "License");
- You may not use this file except in compliance with the License.
- You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+''' added stuck_at column to rules '''
 
- Authors:
- - Martin Barisits, <martin.barisits@cern.ch>, 2014
- - Vincent Garonne, <vincent.garonne@cern.ch>, 2017
-
-Added stuck_at column to rules
-
-Revision ID: 102efcf145f4
-Revises: 4207be2fd914
-Create Date: 2014-10-07 13:31:25.347076
-'''
-
-from alembic.op import add_column, drop_column
 import sqlalchemy as sa
 
-# revision identifiers, used by Alembic.
-revision = '102efcf145f4'  # pylint: disable=invalid-name
-down_revision = '70587619328'  # pylint: disable=invalid-name
+from alembic import context
+from alembic.op import add_column, drop_column
+
+
+# Alembic revision identifiers
+revision = '102efcf145f4'
+down_revision = '70587619328'
 
 
 def upgrade():
     '''
-    upgrade method
+    Upgrade the database to this revision
     '''
-    add_column('rules', sa.Column('stuck_at', sa.DateTime))
+
+    if context.get_context().dialect.name in ['oracle', 'mysql']:
+        add_column('rules', sa.Column('stuck_at', sa.DateTime))
+
+    elif context.get_context().dialect.name == 'postgresql':
+        pass
 
 
 def downgrade():
     '''
-    downgrade method
+    Downgrade the database to the previous revision
     '''
-    drop_column('rules', 'stuck_at')
+
+    if context.get_context().dialect.name in ['oracle', 'mysql']:
+        drop_column('rules', 'stuck_at')
+
+    elif context.get_context().dialect.name == 'postgresql':
+        pass
