+++ lib/rucio/db/sqla/migrate_repo/versions/16a0aca82e12_create_index_on_table_replicas_path.py
@@ -1,37 +1,51 @@
-"""
-  Copyright European Organization for Nuclear Research (CERN)
+# Copyright 2013-2019 CERN for the benefit of the ATLAS collaboration.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# Authors:
+# - Vincent Garonne <vincent.garonne@cern.ch>, 2014-2017
+# - Mario Lassnig <mario.lassnig@cern.ch>, 2019
 
-  Licensed under the Apache License, Version 2.0 (the "License");
-  You may not use this file except in compliance with the License.
-  You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+''' create index on table replicas(path) '''
 
-  Authors:
-  - Vincent Garonne, <vincent.garonne@cern.ch>, 2014-2017
-
- create index on table replicas(path)
-
-Revision ID: 16a0aca82e12
-Revises: None
-Create Date: 2014-02-20 09:35:24.044458
-
-"""
-
+from alembic import context
 from alembic.op import create_index, drop_index
 
-# revision identifiers, used by Alembic.
-revision = '16a0aca82e12'  # pylint:disable=invalid-name
-down_revision = None  # pylint:disable=invalid-name
+
+# Alembic revision identifiers
+revision = '16a0aca82e12'
+down_revision = None
 
 
 def upgrade():
     '''
-    upgrade method
+    Upgrade the database to this revision
     '''
-    create_index('REPLICAS_PATH_IDX', 'replicas', ['path'])
+
+    if context.get_context().dialect.name in ['oracle', 'mysql']:
+        create_index('REPLICAS_PATH_IDX', 'replicas', ['path'])
+
+    elif context.get_context().dialect.name == 'postgresql':
+        pass
 
 
 def downgrade():
     '''
-    downgrade method
+    Downgrade the database to the previous revision
     '''
-    drop_index('REPLICAS_PATH_IDX', 'replicas')
+
+    if context.get_context().dialect.name in ['oracle', 'mysql']:
+        drop_index('REPLICAS_PATH_IDX', 'replicas')
+
+    elif context.get_context().dialect.name == 'postgresql':
+        pass
