+++ lib/rucio/db/sqla/migrate_repo/versions/1a29d6a9504c_add_didtype_chck_to_requests.py
@@ -1,46 +1,59 @@
-# Copyright European Organization for Nuclear Research (CERN)
+# Copyright 2013-2019 CERN for the benefit of the ATLAS collaboration.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
-# You may not use this file except in compliance with the License.
-# You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 #
 # Authors:
-# - Mario Lassnig, <mario.lassnig@cern.ch>, 2014
-# - Vincent Garonne, <vincent.garonne@cern.ch>, 2017
+# - Mario Lassnig <mario.lassnig@cern.ch>, 2014-2019
+# - Vincent Garonne <vincent.garonne@cern.ch>, 2017
 
-"""add didtype_chck to requests
+''' add didtype_chck to requests '''
 
-Revision ID: 1a29d6a9504c
-Revises: 436827b13f82
-Create Date: 2014-10-13 13:48:56.080599
-
-"""
-
-from alembic.op import add_column, drop_column
 import sqlalchemy as sa
+
+from alembic import context
+from alembic.op import add_column, drop_column
 
 from rucio.db.sqla.constants import DIDType
 
-# revision identifiers, used by Alembic.
+# Alembic revision identifiers
 revision = '1a29d6a9504c'
 down_revision = '436827b13f82'
 
 
 def upgrade():
     '''
-    upgrade method
+    Upgrade the database to this revision
     '''
-    add_column('requests', sa.Column('did_type',
-                                     DIDType.db_type(name='REQUESTS_DIDTYPE_CHK'),
-                                     default=DIDType.FILE))
 
-    # we don't want checks on the history table
-    add_column('requests_history', sa.Column('did_type', sa.String(1)))
+    if context.get_context().dialect.name in ['oracle', 'mysql']:
+        add_column('requests', sa.Column('did_type',
+                                         DIDType.db_type(name='REQUESTS_DIDTYPE_CHK'),
+                                         default=DIDType.FILE))
+        # we don't want checks on the history table, fake the DID type
+        add_column('requests_history', sa.Column('did_type', sa.String(1)))
+
+    elif context.get_context().dialect.name == 'postgresql':
+        pass
 
 
 def downgrade():
     '''
-    downgrade method
+    Downgrade the database to the previous revision
     '''
-    drop_column('requests', 'did_type')
-    drop_column('requests_history', 'did_type')
+
+    if context.get_context().dialect.name in ['oracle', 'mysql']:
+        drop_column('requests', 'did_type')
+        drop_column('requests_history', 'did_type')
+
+    elif context.get_context().dialect.name == 'postgresql':
+        pass
