+++ lib/rucio/db/sqla/migrate_repo/versions/4783c1f49cb4_create_distance_table.py
@@ -1,47 +1,55 @@
-# Copyright European Organization for Nuclear Research (CERN)
+# Copyright 2013-2019 CERN for the benefit of the ATLAS collaboration.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
-# You may not use this file except in compliance with the License.
-# You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 #
 # Authors:
-# - Wen Guan, <wen.guan@cern.ch>, 2015
-# - Vincent Garonne, <vincent.garonne@cern.ch>, 2017
+# - Wen Guan <wen.guan@cern.ch>, 2015
+# - Vincent Garonne <vincent.garonne@cern.ch>, 2017
+# - Mario Lassnig <mario.lassnig@cern.ch>, 2019
 
-"""create distance table
+''' create distance table '''
 
-Revision ID: 4783c1f49cb4
-Revises: 277b5fbb41d3
-Create Date: 2015-05-21 08:11:14.318464
+import datetime
 
-"""
+import sqlalchemy as sa
 
+from alembic import context
 from alembic.op import (create_table, create_primary_key, create_foreign_key,
                         create_check_constraint, create_index, drop_table)
-from alembic import context
-import sqlalchemy as sa
 
 from rucio.db.sqla.types import GUID
 
-# revision identifiers, used by Alembic.
+
+# Alembic revision identifiers
 revision = '4783c1f49cb4'
 down_revision = '277b5fbb41d3'
 
 
 def upgrade():
     '''
-    upgrade method
+    Upgrade the database to this revision
     '''
-    create_table('distances',
-                 sa.Column('src_rse_id', GUID()),
-                 sa.Column('dest_rse_id', GUID()),
-                 sa.Column('ranking', sa.Integer),
-                 sa.Column('agis_distance', sa.Integer),
-                 sa.Column('geoip_distance', sa.Integer),
-                 sa.Column('updated_at', sa.DateTime),
-                 sa.Column('created_at', sa.DateTime))
 
-    if context.get_context().dialect.name != 'sqlite':
+    if context.get_context().dialect.name in ['oracle', 'mysql']:
+        create_table('distances',
+                     sa.Column('src_rse_id', GUID()),
+                     sa.Column('dest_rse_id', GUID()),
+                     sa.Column('ranking', sa.Integer),
+                     sa.Column('agis_distance', sa.Integer),
+                     sa.Column('geoip_distance', sa.Integer),
+                     sa.Column('created_at', sa.DateTime, default=datetime.datetime.utcnow),
+                     sa.Column('updated_at', sa.DateTime, default=datetime.datetime.utcnow, onupdate=datetime.datetime.utcnow))
+
         create_primary_key('DISTANCES_PK', 'distances', ['src_rse_id', 'dest_rse_id'])
         create_foreign_key('DISTANCES_SRC_RSES_FK', 'distances', 'rses', ['src_rse_id'], ['id'])
         create_foreign_key('DISTANCES_DEST_RSES_FK', 'distances', 'rses', ['dest_rse_id'], ['id'])
@@ -49,9 +57,17 @@ def upgrade():
         create_check_constraint('DISTANCES_UPDATED_NN', 'distances', 'updated_at is not null')
         create_index('DISTANCES_DEST_RSEID_IDX', 'distances', ['dest_rse_id'])
 
+    elif context.get_context().dialect.name == 'postgresql':
+        pass
+
 
 def downgrade():
     '''
-    downgrade method
+    Downgrade the database to the previous revision
     '''
-    drop_table('distances')
+
+    if context.get_context().dialect.name in ['oracle', 'mysql']:
+        drop_table('distances')
+
+    elif context.get_context().dialect.name == 'postgresql':
+        pass
