+++ lib/rucio/db/sqla/migrate_repo/versions/4bab9edd01fc_create_index_on_requests_rule_id.py
@@ -1,38 +1,52 @@
-"""
-  Copyright European Organization for Nuclear Research (CERN)
+# Copyright 2013-2019 CERN for the benefit of the ATLAS collaboration.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+# Authors:
+# - Wen Guan <wen.guan@cern.ch>, 2015
+# - Vincent Garonne <vincent.garonne@cern.ch>, 2017
+# - Mario Lassnig <mario.lassnig@cern.ch>, 2019
 
-  Licensed under the Apache License, Version 2.0 (the "License");
-  You may not use this file except in compliance with the License.
-  You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+''' create index on requests.rule_id '''
 
-  Authors:
-  - Wen Guan, <wen.guan@cern.ch>, 2015
-  - Vincent Garonne, <vincent.garonne@cern.ch>, 2017
-
-create index on requests.rule_id
-
-Revision ID: 4bab9edd01fc
-Revises: ae2a56fcc89
-Create Date: 2015-03-20 15:58:02.456873
-
-"""
-
+from alembic import context
 from alembic.op import create_index, drop_index
 
-# revision identifiers, used by Alembic.
-revision = '4bab9edd01fc'  # pylint:disable=invalid-name
-down_revision = 'ae2a56fcc89'  # pylint:disable=invalid-name
+
+# Alembic revision identifiers
+revision = '4bab9edd01fc'
+down_revision = 'ae2a56fcc89'
 
 
 def upgrade():
     '''
-    upgrade method
+    Upgrade the database to this revision
     '''
-    create_index('REQUESTS_RULEID_IDX', 'requests', ['rule_id'])
+
+    if context.get_context().dialect.name in ['oracle', 'mysql']:
+        create_index('REQUESTS_RULEID_IDX', 'requests', ['rule_id'])
+
+    elif context.get_context().dialect.name == 'postgresql':
+        pass
 
 
 def downgrade():
     '''
-    downgrade method
+    Downgrade the database to the previous revision
     '''
-    drop_index('REQUESTS_RULEID_IDX', 'requests')
+
+    if context.get_context().dialect.name in ['oracle', 'mysql']:
+        drop_index('REQUESTS_RULEID_IDX', 'requests')
+
+    elif context.get_context().dialect.name == 'postgresql':
+        pass
